--- src/App.css
@@ -1,9 +1,0 @@
-.App {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-  font-family: 'Rubik', sans-serif;
-}
-
