+++ src/components/App.js
@@ -5,9 +5,8 @@ import {
   Link,
   Switch
 } from 'react-router-dom';
-import Homecard from './components/card/home-card';
-import ChartCard from './components/card/chart-card';
-import './App.css';
+import Home from './cards/home/home';
+import Stats from './cards/stats/stats';
 
 const NoMatch = ({ location }) => (
   <div>
@@ -15,17 +14,17 @@ const NoMatch = ({ location }) => (
   </div>
 )
 
-
 class App extends Component {
   render() {
     return (
         <Router>
           <div className="App">
           <Switch>
-            <Route path="/" exact component={Homecard}/>
-            <Route path="/stats" component={ChartCard}/>
+            <Route path="/" exact component={Home}/>
+            <Route path="/stats" component={Stats}/>
             <Route component={NoMatch}/>
           </Switch>
+
           <ul>
             <li><Link to="/">Home</Link></li>
             <li><Link to="/stats">Stats</Link></li>
