+++ src/components/cards/home/home.css
@@ -16,13 +16,15 @@
   align-items: center;
   flex-direction: column;
 }
+
 .card-logo {
   margin-top: 40px;
 }
+
 .header {
   width: 100%;
-display: flex;
-justify-content: center;
+  display: flex;
+  justify-content: center;
 }
 .content{
   align-self: center;
@@ -30,8 +32,8 @@ justify-content: center;
 }
 .footer {
   display: flex;
-align-items: center;
-width: 100%;
-justify-content: space-around;
-height: 100%;
+  align-items: center;
+  width: 100%;
+  justify-content: space-around;
+  height: 100%;
 }
