+++ src/components/cards/stats/stats.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
-import './card.css';
-
-class ChartCard extends Component {
+import '../home/home.css'; // TODO: use proper style
+//  http https://wakatime.com/api/v1/users/current/stats/last_7_days  -a '99c1f1de-2954-41d7-b76e-82cedf9c3fe2'
+class Stats extends Component {
   render() {
     return (
      <div className="card">
@@ -17,4 +17,4 @@ class ChartCard extends Component {
   }
 }
 
-export default ChartCard;
+export default Stats;
