+++ src/index.css
@@ -1,3 +1,12 @@
+.App {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 100%;
+  font-family: 'Rubik', sans-serif;
+}
+
 body {
   margin: 0;
   padding: 0;
